move group_by_window serialization tests to kernel tests
